% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_links}
\alias{get_links}
\title{Crawl Webpage for all links. Requires installation of lynx Browser}
\usage{
get_links(url)
}
\arguments{
\item{url}{A full URL in the form of a string.}
}
\description{
This function returns a tibble of linked URLs from a source URL.
}
\examples{
links <- get_links("https://cnn.com")
}
\keyword{weblinks}
